// packages/shared/src/types.ts
export type UserRole = 'admin' | 'teacher' | 'parent' | 'student'
export type UserStatus = 'active' | 'inactive'

export interface User {
  id: string
  email?: string // Optional for students, required for others
  displayName: string
  role: UserRole
  avatarId: string | null
  status: UserStatus
  emailVerified: boolean
  createdAt: string
  updatedAt: string
}